requirements.txt
python_flickr
